print number sixth commit

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -5,7 +5,7 @@
  * @n: input number
  * Return: None
  */
-void print_number(int n)
+void print_number(long int n)
 {
 	if (n < 0)
 	{
